native: long -> s32 in LP64 signature mismatches (group A)
GC 'long' == 32-bit; on LP64 these decls no longer matched their s32/.cpp uses:
- CardSave changeMode/setMessage; JDRLighting setLightNum/setAmbNum;
  MapObjGrass entryGrassGroup; NpcBase setBalloonMessage;
  MapObjBase loadHideObjInfo (long*->s32*); ToolData GetValue (long&->s32&);
  CameraDemo startDemoCamera arg; dolphin/dvd.h callback first param (long->s32).

+++ include/JSystem/JDrama/JDRLighting.hpp
@@ -98,7 +98,7 @@ public:
 	virtual TNameRef* searchF(u16, const char*);
 	virtual void perform(u32, TGraphics*);
 
-	void setLightNum(long);
+	void setLightNum(s32);
 
 public:
 	/* 0x10 */ TIdxLight* mLights;
@@ -136,7 +136,7 @@ public:
 	virtual TNameRef* searchF(u16, const char*);
 	virtual void perform(u32, TGraphics*) { }
 
-	void setAmbNum(long);
+	void setAmbNum(s32);
 
 public:
 	/* 0x10 */ TAmbColor* mAmbColors;
