+++ include/MoveBG/MapObjBase.hpp
@@ -317,8 +317,8 @@ public:
 	void throwObjToFront(TMapObjBase*, f32, f32, f32) const;
 	void throwObjToOverhead(TMapObjBase*, f32, f32) const;
 	void checkOnManhole();
-	static void loadHideObjInfo(JSUMemoryInputStream&, long*, f32*, f32*,
-	                            long*);
+	static void loadHideObjInfo(JSUMemoryInputStream&, s32*, f32*, f32*,
+	                            s32*);
 	static bool isDemo();
 	static bool isHideObj(THitActor*);
 	void getObjCollisionHeightOffset() const { }
